pyproject.toml
[project]
name = "django-ninja-errors"
version = "0.1.0"
description = ""
authors = [
    {name = "Dima Rubets",email = "rubetsdima100@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-ninja-extra (>=0.22.6,<0.23.0)",
    "pyhumps (>=3.8.0,<4.0.0)",
    "django-phonenumber-field[phonenumberslite] (>=8.0.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
